rename UserActivityLogPopover to Notifications
