+++ src/components/common/Layout/components/Account/components/Notifications/Notifications.tsx
@@ -1,6 +1,5 @@
 import {
   Divider,
-  HStack,
   Icon,
   Popover,
   PopoverArrow,
@@ -8,34 +7,24 @@ import {
   PopoverCloseButton,
   PopoverContent,
   PopoverTrigger,
-  Skeleton,
-  SkeletonCircle,
   Stack,
-  Text,
   useColorModeValue,
 } from "@chakra-ui/react"
-import { ActivityLogActionProvider } from "components/[guild]/activity/ActivityLogAction/ActivityLogActionContext"
-import ActionIcon from "components/[guild]/activity/ActivityLogAction/components/ActionIcon"
-import ActionLabel from "components/[guild]/activity/ActivityLogAction/components/ActionLabel"
-import {
-  ActivityLogProvider,
-  useActivityLog,
-} from "components/[guild]/activity/ActivityLogContext"
+import { ActivityLogProvider } from "components/[guild]/activity/ActivityLogContext"
 import useUser from "components/[guild]/hooks/useUser"
 import { usePostHogContext } from "components/_app/PostHogProvider"
 import useWeb3ConnectionManager from "components/_app/Web3ConnectionManager/hooks/useWeb3ConnectionManager"
-import LinkButton from "components/common/LinkButton"
 import useLocalStorage from "hooks/useLocalStorage"
 import dynamic from "next/dynamic"
-import { ArrowRight, Bell } from "phosphor-react"
-import GhostIcon from "static/avatars/ghost.svg"
-import AccountButton from "./AccountButton"
-import NotificationsSection from "./NotificationsSection"
+import { Bell } from "phosphor-react"
+import AccountButton from "../AccountButton"
+import NotificationsActivityLog from "./components/NotificationsActivityLog"
+import NotificationsSection from "./components/NotificationsSection"
 
-const DynamicWeb3Inbox = dynamic(() => import("./Web3Inbox"))
+const DynamicWeb3Inbox = dynamic(() => import("../Web3Inbox"))
 const VIEWPORT_GAP_PX = 8
 
-const UserActivityLogPopover = () => {
+const Notifications = () => {
   const { id } = useUser()
   const { captureEvent } = usePostHogContext()
 
@@ -65,7 +54,7 @@ const UserActivityLogPopover = () => {
         <>
           <PopoverTrigger>
             <AccountButton
-              aria-label="Activity log"
+              aria-label="Notifications"
               onClick={() => {
                 setClickedOnNotifications(true)
                 if (isOpen) return
@@ -117,7 +106,7 @@ const UserActivityLogPopover = () => {
                   isInfinite={false}
                 >
                   <NotificationsSection title="Recent activity">
-                    <UserActivityLog />
+                    <NotificationsActivityLog />
                   </NotificationsSection>
                 </ActivityLogProvider>
               </Stack>
@@ -129,62 +118,4 @@ const UserActivityLogPopover = () => {
   )
 }
 
-const SHOWN_ACTIONS_COUNT = 3
-const UserActivityLog = (): JSX.Element => {
-  const { data, isValidating } = useActivityLog()
-
-  return (
-    <>
-      <Stack minW="xs" spacing={0.5} divider={<Divider />}>
-        {!data || isValidating ? (
-          [...Array(SHOWN_ACTIONS_COUNT)].map((_, i) => (
-            <HStack key={i} spacing={4} py={4}>
-              <SkeletonCircle boxSize={8} />
-              <Stack spacing={1}>
-                <Skeleton h={5} w={36} />
-                <Skeleton h={4} w={20} />
-              </Stack>
-            </HStack>
-          ))
-        ) : !data.entries.length ? (
-          <HStack py="3" spacing="3.5">
-            <Icon as={GhostIcon} boxSize={5} alt="Not found" />
-            <Text>No actions yet</Text>
-          </HStack>
-        ) : (
-          data.entries.slice(0, SHOWN_ACTIONS_COUNT)?.map((action) => (
-            <ActivityLogActionProvider key={action.id} action={action}>
-              <HStack spacing={4} py={4} pl="1">
-                <ActionIcon size={6} />
-                <Stack spacing={{ base: 1, md: 0.5 }}>
-                  <ActionLabel />
-                  <Text as="span" colorScheme="gray" fontSize="sm">
-                    {new Date(Number(action.timestamp)).toLocaleString()}
-                    {/* TODO: would be nice to display user friendly "Today at x" or "x days ago" relative times */}
-                    {/* {`${formatRelativeTimeFromNow(
-                      Date.now() - parseInt(action.timestamp)
-                    )} ago`} */}
-                  </Text>
-                </Stack>
-              </HStack>
-            </ActivityLogActionProvider>
-          ))
-        )}
-      </Stack>
-      {data?.entries?.length > 0 && (
-        <LinkButton
-          href="/profile/activity"
-          w="full"
-          h="10"
-          mt="2"
-          colorScheme="gray"
-          rightIcon={<Icon as={ArrowRight} />}
-        >
-          View all
-        </LinkButton>
-      )}
-    </>
-  )
-}
-
-export default UserActivityLogPopover
+export default Notifications
