+++ src/components/common/Layout/components/Account/components/Notifications/index.ts
@@ -1,0 +1,3 @@
+import Notifications from "./Notifications"
+
+export default Notifications
